requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-genai==0.1.0
pillow==10.1.0
pydantic==2.5.0